move prompt types to voicePrompt
